backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
openai>=1.30.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
